add: delete pembayaran

+++ app/Http/Controllers/Ramwater/Penjualan/PembayaranController.php
@@ -9,10 +9,12 @@ use App\Models\DPelangganModel;
 use App\Models\DPembayaranModel;
 use App\Models\DPenjualanModel;
 use App\Models\Karyawan;
+use App\Models\Penjualan;
 use App\Models\TChannelModel;
 use App\Models\TGudang;
 use App\Services\PembayaranService;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Str;
 
 class PembayaranController extends Controller
@@ -25,6 +27,7 @@ class PembayaranController extends Controller
     {
         $this->integrationHelper = new IntegrationHelper();
         $this->dPembayaranModel = new DPembayaranModel();
+        $this->dPenjualanModel = new Penjualan();
         $this->pembayaranService = $pembayaranService;
     }
 
@@ -64,48 +67,37 @@ class PembayaranController extends Controller
 
     public function destroy($id)
     {
-        $pembayaran = $this->dPembayaranModel->where('id', '=', $id)->first();
-        $penjualan = $this->dPenjualanModel->where('nota_penjualan', '=', $pembayaran->nota_penjualan)->first();
-        $bayarDetails = $this->dPembayaranModel->where('nota_penjualan', '=', $pembayaran->nota_penjualan)->get();
+        return DB::transaction(function () use ($id) {
+            $pembayaran = $this->dPembayaranModel->where('id', '=', $id)->first();
+            if (!$pembayaran) return response()->json(['success' => false, 'message' => 'pembayaran tidak ditemukan']);
 
-        if ($pembayaran->path_file) {
-            // hapus d_penjualan & child
-            $pathToDelete = $pembayaran->path_file;
-            $publicPath = storage_path('app/public/');
+            $penjualan = $this->dPenjualanModel->where('nota_penjualan', '=', $pembayaran->nota_penjualan)->first();
 
-            // Pastikan path_file dimulai dengan "storage/"
-            if (Str::startsWith($pathToDelete, 'storage/')) {
-                // Ubah "storage/" menjadi direktori penyimpanan
-                $pathToDelete = $publicPath . Str::after($pathToDelete, 'storage/');
+            if ($pembayaran->path_file) {
+                // hapus d_penjualan & child
+                $pathToDelete = $pembayaran->path_file;
+                $publicPath = storage_path('app/public/');
+
+                // Pastikan path_file dimulai dengan "storage/"
+                if (Str::startsWith($pathToDelete, 'storage/')) {
+                    // Ubah "storage/" menjadi direktori penyimpanan
+                    $pathToDelete = $publicPath . Str::after($pathToDelete, 'storage/');
+                }
+
+                FormatHelper::deleteFile($pathToDelete);
             }
+            $pembayaran->delete();
 
-            FormatHelper::deleteFile($pathToDelete);
-        }
-        $pembayaran->delete();
+            $penjualanModel = new Penjualan();
+            $penjualan = $penjualanModel->find($penjualan->id);
 
-        $penjualanModel = new DPenjualanModel(); // Gantilah Penjualan dengan nama model yang sesuai
-        $penjualan = $penjualanModel->find($penjualan->id);
+            $totalNominalBayar = new DPembayaranModel();
+            $totalNominalBayar = $totalNominalBayar->where('nota_penjualan', '=', $penjualan->nota_penjualan)->get();
+            $totalNominalBayar = $totalNominalBayar->sum('nominal_bayar');
+            $this->pembayaranService->updateStatus($penjualan, $totalNominalBayar, $penjualan->harga_total);
 
-        $totalNominalBayar = new DPembayaranModel();
-        $totalNominalBayar = $totalNominalBayar->where('nota_penjualan', '=', $penjualan->nota_penjualan)->get();
-        $totalNominalBayar = $totalNominalBayar->sum('nominal_bayar');
-
-        if ($totalNominalBayar == $penjualan->harga_total) {
-            $penjualan->sts_angsuran = 4;
-            $penjualan->nominal_bayar = $totalNominalBayar;
-            $penjualan->sisa_bayar = $penjualan->harga_total - $totalNominalBayar;
-            $penjualan->save();
-        } elseif ($totalNominalBayar < $penjualan->harga_total) {
-            $penjualan->sts_angsuran = 1;
-            $penjualan->nominal_bayar = $totalNominalBayar;
-            $penjualan->sisa_bayar = $penjualan->harga_total - $totalNominalBayar;
-            $penjualan->save();
-        } elseif ($totalNominalBayar > $penjualan->harga_total) {
-            throw new \Exception("Pembayaran Terlalu banyak!");
-        };
-
-
-        return response()->json(['success' => true, 'message' => 'Data berhasil dihapus']);
+            return response()->json(['success' => true, 'message' => 'Data berhasil dihapus']);
+        });
     }
 
     public function laporan()
